+++ st2common/tests/unit/test_queue_consumer.py
@@ -14,10 +14,10 @@
 # limitations under the License.
 
 import mock
-from kombu import Connection, Exchange, Queue
+from kombu import Exchange, Queue
 
 from st2common.transport import consumers
-from st2common.transport import utils as transport_utils
+from st2common.util.greenpooldispatch import BufferedDispatcher
 from st2tests.base import DbTestCase
 from tests.unit.base import FakeModelDB
 
@@ -34,8 +34,7 @@ class FakeMessageHandler(consumers.MessageHandler):
 
 
 def get_handler():
-    with Connection(transport_utils.get_messaging_urls()) as conn:
-        return FakeMessageHandler(conn, [FAKE_WORK_Q])
+    return FakeMessageHandler(mock.MagicMock(), [FAKE_WORK_Q])
 
 
 class QueueConsumerTest(DbTestCase):
@@ -51,5 +50,53 @@ class QueueConsumerTest(DbTestCase):
     def test_process_message_wrong_payload_type(self):
         payload = 100
         handler = get_handler()
-        handler._queue_consumer._process_message(payload)
+        mock_message = mock.MagicMock()
+        handler._queue_consumer.process(payload, mock_message)
+        self.assertTrue(mock_message.ack.called)
         self.assertFalse(FakeMessageHandler.process.called)
+
+
+class FakeStagedMessageHandler(consumers.StagedMessageHandler):
+    message_type = FakeModelDB
+
+    def pre_ack_process(self, message):
+        return message
+
+    def process(self, payload):
+        pass
+
+
+def get_staged_handler():
+    return FakeStagedMessageHandler(mock.MagicMock(), [FAKE_WORK_Q])
+
+
+class StagedQueueConsumerTest(DbTestCase):
+
+    @mock.patch.object(FakeStagedMessageHandler, 'pre_ack_process', mock.MagicMock())
+    def test_process_message_pre_ack(self):
+        payload = FakeModelDB()
+        handler = get_staged_handler()
+        mock_message = mock.MagicMock()
+        handler._queue_consumer.process(payload, mock_message)
+        FakeStagedMessageHandler.pre_ack_process.assert_called_once_with(payload)
+        self.assertTrue(mock_message.ack.called)
+
+    @mock.patch.object(BufferedDispatcher, 'dispatch', mock.MagicMock())
+    @mock.patch.object(FakeStagedMessageHandler, 'process', mock.MagicMock())
+    def test_process_message(self):
+        payload = FakeModelDB()
+        handler = get_staged_handler()
+        mock_message = mock.MagicMock()
+        handler._queue_consumer.process(payload, mock_message)
+        BufferedDispatcher.dispatch.assert_called_once_with(
+            handler._queue_consumer._process_message, payload)
+        handler._queue_consumer._process_message(payload)
+        FakeStagedMessageHandler.process.assert_called_once_with(payload)
+        self.assertTrue(mock_message.ack.called)
+
+    def test_process_message_wrong_payload_type(self):
+        payload = 100
+        handler = get_staged_handler()
+        mock_message = mock.MagicMock()
+        handler._queue_consumer.process(payload, mock_message)
+        self.assertTrue(mock_message.ack.called)