+++ CHANGELOG.rst
@@ -5,7 +5,9 @@ in development
 --------------
 
 * SSL support for mongodb connections. (improvement)
-
+* TriggerInstances now have statuses to help track if a TriggerInstance has been processed,
+  is being processed or failed to process. This bring out some visibility into parts of the
+  TriggerInstance processing pipeline and can help identify missed events. (new-feature)
 
 1.4.0 - April 18, 2016
 ----------------------